+++ src/entity/ec_first_person_camera.cpp
@@ -9,12 +9,12 @@ namespace entity {
     }
 
     void ECFirstPersonCamera::position_camera() {
-        auto &entity_transform = this->entity.get<Transform>();
+        auto &entity_transform = this->entity.get<ECTransform>();
         global.game->camera->transform.pos = entity_transform.pos;
     }
 
     void ECFirstPersonCamera::on_position_change(PositionChangeEvent event) {
-        auto &entity_transform = this->entity.get<Transform>();
+        auto &entity_transform = this->entity.get<ECTransform>();
         global.game->camera->transform.pos = entity_transform.pos;
     }
