+++ src/entity/move_system.cpp
@@ -5,7 +5,7 @@
 namespace entity {
     void MoveSystem::update() {
         for (auto &entity : entities) {
-            auto &transform = entity.get<Transform>();
+            auto &transform = entity.get<ECTransform>();
 
             transform.rotate(1, { 0, 1, 0 });
         }
